Refactor code and implement new wrapper component
Before the commit, the product list would handle its functionality to
slide and fade into view. It would be much useful to separate this
concern into its own wrapper component, as this better modularizes
the product list and provides reusability for this functionality. Now,
the code is refactored, and the wrapper component is wrapping elements
wherever its functionality is needed.